backend/requirements.txt
boto3==1.40.55
python-dotenv==1.1.1
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0